requirements.txt
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
Flask==2.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mercadopago==2.1.0
Pillow==9.1.1
pycparser==2.21
pyOpenSSL==22.0.0
pyTelegramBotAPI==4.19.1
python-dotenv==1.0.1
qrcode==7.3.1
requests==2.28.0
requests-pkcs12==1.14
telebot==0.0.5
urllib3==1.26.9
Werkzeug==2.1.2
